+++ inst/examples/any-layer.R
@@ -14,7 +14,7 @@ properties <- list(
   wireframe = TRUE,
   lineWidthMinPixels = 1,
   getPolygon = JS("d => d.contour"),
-  getElevation = JS ("d => d.population / d.area / 10"),
+  getElevation = JS("d => d.population / d.area / 10"),
   getFillColor = JS("d => [d.population / d.area / 60, 140, 0]"),
   getLineColor = c(80, 80, 80),
   getLineWidth = 1