+++ inst/examples/deckgl-api-example_icon-layer.R
@@ -22,7 +22,7 @@ properties <- list(
   iconMapping = list(marker = icon_definition()),
   sizeScale = 15,
   getPosition = get_position(coordinates = "coordinates"), # JS("d => d.coordinates"),
-  #getPosition = get_position(latitude = "lat", longitude = "lng"),
+  # getPosition = get_position(latitude = "lat", longitude = "lng"),
   getIcon = JS("d => 'marker'"),
   getSize = 5,
   getColor = JS("d => [Math.sqrt(d.exits), 140, 0]"),